Refactored asserters into lib & test added
